+++ include/fdreadoutlibs/crt/CRTFrameProcessor.hpp
@@ -6,15 +6,19 @@
  * received with this code.
  */
 #ifndef FDREADOUTLIBS_INCLUDE_FDREADOUTLIBS_CRT_CRTFRAMEPROCESSOR_HPP_
-#define FDREADOUTLIBS_INCLUDE_FDREADOUTLIBS_CRT_CRTFRAMEPROCESSOR_HPP_
+#define FDREADOUTLIBS_INCLUDE_FDREADOUTLIBS_CRT_CRTFIIXEDSIZEFRAMEPROCESSOR_HPP_
 
+#include "logging/Logging.hpp"
+
+#include "readoutlibs/FrameErrorRegistry.hpp"
 #include "readoutlibs/ReadoutIssues.hpp"
+#include "readoutlibs/ReadoutLogging.hpp"
 #include "readoutlibs/models/TaskRawDataProcessorModel.hpp"
 
 #include "fddetdataformats/CRTFrame.hpp"
-#include "logging/Logging.hpp"
+
 #include "fdreadoutlibs/CRTTypeAdapter.hpp"
-#include "readoutlibs/ReadoutLogging.hpp"
+
 
 #include <atomic>
 #include <functional>
@@ -22,35 +26,35 @@
 #include <string>
 
 using dunedaq::readoutlibs::logging::TLVL_BOOKKEEPING;
-using dunedaq::readoutlibs::logging::TLVL_FRAME_RECEIVED;
 
 namespace dunedaq {
 namespace fdreadoutlibs {
 
 class CRTFrameProcessor : public readoutlibs::TaskRawDataProcessorModel<types::CRTTypeAdapter>
 {
+
 public:
   using inherited = readoutlibs::TaskRawDataProcessorModel<types::CRTTypeAdapter>;
   using frameptr = types::CRTTypeAdapter*;
   using crtframeptr = dunedaq::fddetdataformats::CRTFrame*;
   using timestamp_t = std::uint64_t; // NOLINT(build/unsigned)
 
+  // Constructor
   explicit CRTFrameProcessor(std::unique_ptr<readoutlibs::FrameErrorRegistry>& error_registry)
     : readoutlibs::TaskRawDataProcessorModel<types::CRTTypeAdapter>(error_registry)
   {}
 
-  // Custom pipeline registration
+  // Override config for pipeline setup
   void conf(const nlohmann::json& args) override;
 
 protected:
-
   /**
-   * Pipeline Stage 1.: Check proper timestamp increments in DAPHNE frame
+   * Pipeline Stage 1.: Check proper timestamp increments in CRT frame
    * */
   void timestamp_check(frameptr /*fp*/);
 
   /**
-   * Pipeline Stage 2.: Check headers for error flags
+   * Pipeline Stage 2.: Check CRT headers for error flags
    * */
   void frame_error_check(frameptr /*fp*/);
 
@@ -61,11 +65,11 @@ protected:
   bool m_first_ts_missmatch = true;
   bool m_problem_reported = false;
   std::atomic<int> m_ts_error_ctr{ 0 };
+
 private:
 };
 
-} // namespace ndreadoutlibs
+} // namespace fdreadoutlibs
 } // namespace dunedaq
-
 
 #endif // FDREADOUTLIBS_INCLUDE_FDREADOUTLIBS_CRT_CRTFRAMEPROCESSOR_HPP_
